Upgrade to use a proper screen
This is considered a major update to the MakeIcon project. It not only adds a proper screen for previewing the selected images to be used for the icon but also changes the style of the icon of the application.

However, the biggest change in this version is the change from in license from MIT to Apache. Despite being more verbose, it clearly states that any contribution to the project goes under the same license, which could be a source of misunderstanding with the MIT. Details of the reason can be read on the NOTICE file.